US62T70 Move projects endpoint tests into file
